feat: improve TransactionResponse
